adds quiz submission for new quiz style and only one result on results page
