+++ src/quiz/aura/ResultsPage.css
@@ -77,4 +77,4 @@
     text-align: center;
     margin-top: 50px;
   }
-  
+
